refacted db and images
