+++ geopaysagesftpclient/geopaysagesftpclient/__init__.py
@@ -1,17 +1,9 @@
 import os
 import sys
-
 from configparser import ConfigParser
-from iptcinfo3 import IPTCInfo, c_datasets
-from PIL import Image, ImageFile, IptcImagePlugin
-from PIL.IptcImagePlugin import getiptcinfo
-from sqlalchemy import engine_from_config, text
-from sqlalchemy.engine import Engine
 
 from .client import gpClient
 from .patterns import date_from_group_dict
-
-ImageFile.LOAD_TRUNCATED_IMAGES = True
 
 def connect(host:str, port:int=21, user:str=None, pwd:str=None) -> gpClient:
     '''Establishes an ftp connection
@@ -126,126 +118,6 @@ def read_config_file(config_filename:str) -> dict:
 
     except:
         raise Exception('Could not parse config file')
-
-def retrieve_copyright_notice(fn:str) -> dict:
-    '''Retrieves the copyright notice from the input file
-    Returns
-    -------
-    Tuple (copyright notice, description)
-    '''
-    info = IPTCInfo(fn)
-    return {
-        key: info[key]
-        for key in c_datasets.values() if info[key]
-    }
-
-def process_image(ifile:str, siteconfig, ofile:str=None):
-    '''resize an image and return the output file name and the exif and iptc meta data'''
-    try:
-        size = siteconfig.get('resize')
-        site_copyright_notice = siteconfig.get('copyright_notice')
-
-        im  = Image.open(ifile)
-        im_xif = im.info.get('exif')
-        im_iptc = retrieve_copyright_notice(ifile)
-
-        dest = ofile if size and ofile else ifile #Target the input file if the output file is not provided
-        
-        # Image resizing with exif preservation
-        if size:
-            im2 = im.resize(size)
-
-            params = dict(exif=im_xif) if im_xif else dict()   
-            im2.save(dest, **params)
-
-        # IPTC data rewriting. IPTC are lost during the resize operation. We have to put them back manually
-        # If the inputfile contains a copyright notice, then use it instead of the siteconfig value
-        im_iptc['copyright notice'] = im_iptc.get('copyright notice', site_copyright_notice)
-
-        if im_iptc or site_copyright_notice:
-            iptc = IPTCInfo(dest)
-
-            for key in im_iptc:
-                iptc[key] = im_iptc[key]
-            iptc.save()
-
-        return (dest, im_xif, im_iptc)
-    except Exception as e:
-        printfailure('could not process image ', ifile, ' traceback ', str(e))
-        return (ifile, None, None)
-
-def sqlalchemy_engine_from_config(configfile:str) -> Engine:
-    '''Returns an sqlalchemy engine'''
-    with open(configfile, 'r') as cf:
-        config = ConfigParser()
-        config.read_file(cf)
-
-        return engine_from_config(
-            dict(config.items('main')),
-            prefix='sqlalchemy.'
-        )
-
-def get_site_id(engine: Engine, sitename:str):
-    '''Returns a site id'''
-    try:
-        return engine.execute(
-            text('select id_site from geopaysages.t_site where name_site=:name limit 1'),
-            name=sitename
-        ).scalar()
-    except:
-        printfailure('Could not retrieve site : "', sitename, '" id from the database')
-        return None
-
-def get_licence_id (engine: Engine, iptc:dict):
-    '''Get the Licence id that matches the copyright notice or create it if none exists in the database'''
-    if not iptc:
-        return None
-
-    notice = iptc.get('copyright notice')
-    cnx = engine.connect()
-
-    id_licence_photo = cnx.execute(
-        text(
-        'select id_licence_photo from geopaysages.dico_licence_photo where name_licence_photo = :nt'
-        ), nt=notice
-    ).scalar()
-
-    if not id_licence_photo:
-        id_licence_photo = cnx.execute(
-            text(
-                'insert into geopaysages.dico_licence_photo (name_licence_photo, description_licence_photo) values (:nt,:desc) returning id_licence_photo'
-            ), nt=notice, desc=notice
-        ).scalar()
-
-    return id_licence_photo
-
-def insert_image_in_db(engine: Engine, siteid:int, matchdict: dict, exif=None, iptc=None):
-    query = text('insert into geopaysages.t_photo \
-        (id_site, path_file_photo, date_photo, filter_date, display_gal_photo, id_licence_photo)\
-        values (:id_site, :path, :strfdate, :f_date, :display, :id_licence)'
-    )
-
-    id_licence_photo = get_licence_id(engine, iptc) if iptc else None
-    filter_date = date_from_group_dict(matchdict)
-    cnx = engine.connect()
-
-    tran = cnx.begin()
-    try:
-        cnx.execute(
-            query,
-            id_site = siteid,
-            path = matchdict.get('ofilename'),
-            strfdate = date_from_group_dict(matchdict).strftime('%d/%m/%y'),
-            f_date = filter_date,
-            display = True,
-            id_licence=id_licence_photo
-        )
-        tran.commit()
-    except:
-        printfailure('Could not insert image ', matchdict.get('ofilename'), ' into database')
-        tran.rollback()
-        raise
-
 
 def printfailure(*args, **kwargs):
     '''prints error message into stderr for failure tolerent operations'''
